Fix/current user routes

+++ src/main/java/com/weg/weg_skills/service/CourseService.java
@@ -6,6 +6,7 @@ import com.weg.weg_skills.exceptions.ResourceNotFoundException;
 import com.weg.weg_skills.mapper.CourseMapper;
 import com.weg.weg_skills.model.Course;
 import com.weg.weg_skills.repository.CourseRepository;
+import com.weg.weg_skills.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,6 +19,7 @@ public class CourseService {
     private CourseRepository courseRepository;
     private CourseMapper courseMapper;
     private MediaService mediaService;
+    private UserRepository userRepository;
 
     public CourseResponseDTO create(CourseCreateRequestDTO dto) {
         if (courseRepository.existsByTitleIgnoreCase(dto.title())) {
@@ -32,10 +34,14 @@ public class CourseService {
     }
 
     @Transactional
-    public UploadTicketResponseDTO uploadImage(Long id, CreateMediaUploadRequestDTO dto) {
+    public UploadTicketResponseDTO uploadImage(Long id, CreateMediaUploadRequestDTO dto, Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new ResourceNotFoundException("User", userId);
+        }
+
         Course course = courseRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Course", id));
 
-        CreatedMediaUpload createdMedia = mediaService.createCourseImageUpload(course.getId(), null, dto);
+        CreatedMediaUpload createdMedia = mediaService.createCourseImageUpload(course.getId(), userId, dto);
 
         if (course.getImage() != null) {
             mediaService.delete(course.getImage().getId());
